+++ Frontend/lib/main.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:WellCareBot/screens/Authentication/login.dart';
 import 'package:WellCareBot/screens/Home/homepage2.dart';
-import 'package:WellCareBot/screens/Home/settings.dart';
+import 'package:WellCareBot/screens/settings/settings.dart';
 import 'package:WellCareBot/services/shared_preferences.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
