+++ Frontend/lib/screens/booking/booking.dart
@@ -1,5 +1,5 @@
 import 'package:WellCareBot/models/booking_model.dart';
-import 'package:WellCareBot/screens/Home/booking_details.dart';
+import 'package:WellCareBot/screens/booking/booking_details.dart';
 import 'package:WellCareBot/services/cloud_service.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
