+++ Frontend/lib/screens/settings/history.dart
