+++ Frontend/lib/screens/settings/privacy_and_policy.dart
